+++ lib/db.js
@@ -504,6 +504,44 @@ const self = {
 		});
 	},
 
+	get_message: (id, revision) => {
+		return new Promise(async (resolve, reject) => {
+
+			try {
+
+				id = parseInt(id);
+				let query = await self.query(`
+					SELECT message.*,
+						   person.name AS person_name, person.slug AS person_slug,
+						   context.slug AS context_slug
+					FROM message, person, context
+					WHERE message.id = $1
+					  AND message.person_id = person.id
+					  AND message.context_id = context.id
+				`, [id]);
+
+				if (query.rows.length == 0) {
+					// No message found, but we still resolve().
+					return resolve(null);
+				}
+
+				let message = query.rows[0];
+				await self.add_message_details([message]);
+
+				if (revision) {
+					message.revision = revision;
+					message.content = message.revisions[revision].content;
+				}
+
+				resolve(message);
+
+			} catch(err) {
+				console.log(err.stack);
+				reject(err);
+			}
+		});
+	},
+
 	add_facets: (target, target_type, facet_type) => {
 		return new Promise(async (resolve, reject) => {
